Move growth policies in their own namespace (tsl::hh::*) and own file (hopscotch_growth_policy.h).
