+++ src/hopscotch_sc_map.h
@@ -61,7 +61,7 @@ template<class Key,
          class Allocator = std::allocator<std::pair<const Key, T>>,
          unsigned int NeighborhoodSize = 62,
          bool StoreHash = false,
-         class GrowthPolicy = tsl::power_of_two_growth_policy<2>>
+         class GrowthPolicy = tsl::hh::power_of_two_growth_policy<2>>
 class hopscotch_sc_map {
 private:
     template<typename U>
