+++ src/hopscotch_sc_set.h
@@ -60,7 +60,7 @@ template<class Key,
          class Allocator = std::allocator<Key>,
          unsigned int NeighborhoodSize = 62,
          bool StoreHash = false,
-         class GrowthPolicy = tsl::power_of_two_growth_policy<2>>
+         class GrowthPolicy = tsl::hh::power_of_two_growth_policy<2>>
 class hopscotch_sc_set {
 private:    
     template<typename U>
