+++ tests/policy_tests.cpp
@@ -4,171 +4,73 @@
 #include <boost/mpl/list.hpp> 
 #include <cstddef>
 #include <limits>
+#include <ratio>
 #include <stdexcept>
 
-#include "hopscotch_hash.h"
+#include "hopscotch_growth_policy.h"
 
 
 BOOST_AUTO_TEST_SUITE(test_policy)
 
-BOOST_AUTO_TEST_CASE(test_power_of_two_policy) {
-    // Call next_bucket_count() on polcy until we reach its max_bucket_count()
+using test_types = boost::mpl::list<tsl::hh::power_of_two_growth_policy<2>,
+                                    tsl::hh::power_of_two_growth_policy<4>,
+                                    tsl::hh::prime_growth_policy,
+                                    tsl::hh::mod_growth_policy<>,
+                                    tsl::hh::mod_growth_policy<std::ratio<7,2>>>;
+
+
+BOOST_AUTO_TEST_CASE_TEMPLATE(test_policy, Policy, test_types) {
+    // Call next_bucket_count() on the policy until we reach its max_bucket_count()
     bool exception_thrown = false;
     
     std::size_t bucket_count = 0;
-    tsl::power_of_two_growth_policy<2> policy(bucket_count);
-    
-    BOOST_CHECK_EQUAL(policy.bucket_for_hash(0), 0);
-    BOOST_CHECK_EQUAL(bucket_count, 0);
-    
-    
-    try {
-        while(true) {
-            bucket_count = policy.next_bucket_count();
-            policy = tsl::power_of_two_growth_policy<2>(bucket_count);
-            
-            BOOST_CHECK_EQUAL(policy.bucket_for_hash(0), 0);
-            BOOST_CHECK(bucket_count > 0);
-        }
-    }
-    catch(const std::length_error& ) {
-        exception_thrown = true;
-        BOOST_CHECK_EQUAL(bucket_count, policy.max_bucket_count());
-    }
-    
-    BOOST_CHECK(exception_thrown);
-}
-
-BOOST_AUTO_TEST_CASE(test_power_of_two_policy_min_size) {
-    std::size_t size = 0;
-    tsl::power_of_two_growth_policy<2> policy(size);
-    
-    BOOST_CHECK_EQUAL(policy.bucket_for_hash(0), 0);
-}
-
-BOOST_AUTO_TEST_CASE(test_power_of_two_policy_max_size) {
-    std::size_t size = 0;
-    tsl::power_of_two_growth_policy<2> policy(size);
-    
-    
-    size = policy.max_bucket_count();
-    tsl::power_of_two_growth_policy<2> policy2(size);
-    
-    
-    size = std::numeric_limits<std::size_t>::max();
-    BOOST_CHECK_THROW((tsl::power_of_two_growth_policy<2>(size)), std::length_error);
-    
-    
-    size = policy.max_bucket_count() + 1;
-    BOOST_CHECK_THROW((tsl::power_of_two_growth_policy<2>(size)), std::length_error);
-}
-
-
-
-
-BOOST_AUTO_TEST_CASE(test_prime_policy) {
-    // Call next_bucket_count() on polcy until we reach its max_bucket_count()
-    bool exception_thrown = false;
-    
-    std::size_t bucket_count = 0;
-    tsl::prime_growth_policy policy(bucket_count);
-    
-    BOOST_CHECK_EQUAL(policy.bucket_for_hash(0), 0);
-    BOOST_CHECK_EQUAL(bucket_count, 0);
-    
-    
-    try {
-        while(true) {
-            bucket_count = policy.next_bucket_count();
-            policy = tsl::prime_growth_policy(bucket_count);
-            
-            BOOST_CHECK_EQUAL(policy.bucket_for_hash(0), 0);
-            BOOST_CHECK(bucket_count > 0);
-        }
-    }
-    catch(const std::length_error& ) {
-        exception_thrown = true;
-        BOOST_CHECK_EQUAL(bucket_count, policy.max_bucket_count());
-    }
-    
-    BOOST_CHECK(exception_thrown);
-}
-
-BOOST_AUTO_TEST_CASE(test_prime_policy_min_size) {
-    std::size_t size = 0;
-    tsl::prime_growth_policy policy(size);
-    
-    BOOST_CHECK_EQUAL(policy.bucket_for_hash(0), 0);
-}
-
-BOOST_AUTO_TEST_CASE(test_prime_policy_max_size) {
-    std::size_t size = 0;
-    tsl::prime_growth_policy policy(size);
-    
-    
-    size = policy.max_bucket_count();
-    tsl::prime_growth_policy policy2(size);
-    
-    
-    size = std::numeric_limits<std::size_t>::max();
-    BOOST_CHECK_THROW((tsl::prime_growth_policy(size)), std::length_error);
-    
-    
-    size = policy.max_bucket_count() + 1;
-    BOOST_CHECK_THROW((tsl::prime_growth_policy(size)), std::length_error);
-}
-
-
-
-
-BOOST_AUTO_TEST_CASE(test_mod_policy) {
-    // Call next_bucket_count() on polcy until we reach its max_bucket_count()
-    bool exception_thrown = false;
-    
-    std::size_t bucket_count = 0;
-    tsl::mod_growth_policy<> policy(bucket_count);
+    Policy policy(bucket_count);
     
     BOOST_CHECK_EQUAL(policy.bucket_for_hash(0), 0);
     BOOST_CHECK_EQUAL(bucket_count, 0);
     
     try {
         while(true) {
+            const std::size_t previous_bucket_count = bucket_count;
+            
             bucket_count = policy.next_bucket_count();
-            policy = tsl::mod_growth_policy<>(bucket_count);
+            policy = Policy(bucket_count);
             
             BOOST_CHECK_EQUAL(policy.bucket_for_hash(0), 0);
-            BOOST_CHECK(bucket_count > 0);
+            BOOST_CHECK(bucket_count > previous_bucket_count);
         }
     }
     catch(const std::length_error& ) {
         exception_thrown = true;
-        BOOST_CHECK_EQUAL(bucket_count, policy.max_bucket_count());
     }
     
     BOOST_CHECK(exception_thrown);
 }
 
-BOOST_AUTO_TEST_CASE(test_mod_policy_min_size) {
-    std::size_t size = 0;
-    tsl::mod_growth_policy<> policy(size);
+BOOST_AUTO_TEST_CASE_TEMPLATE(test_policy_min_bucket_count, Policy, test_types) {
+    // Check polcy when a bucket_count of 0 is asked.
+    std::size_t bucket_count = 0;
+    Policy policy(bucket_count);
     
     BOOST_CHECK_EQUAL(policy.bucket_for_hash(0), 0);
 }
 
-BOOST_AUTO_TEST_CASE(test_mod_policy_max_size) {
-    std::size_t size = 0;
-    tsl::mod_growth_policy<> policy(size);
-    
-    size = policy.max_bucket_count();
-    tsl::mod_growth_policy<> policy2(size);
+BOOST_AUTO_TEST_CASE_TEMPLATE(test_policy_max_bucket_count, Policy, test_types) {
+    // Test a bucket_count equals to the max_bucket_count limit and above
+    std::size_t bucket_count = 0;
+    Policy policy(bucket_count);
     
     
-    size = std::numeric_limits<std::size_t>::max();
-    BOOST_CHECK_THROW((tsl::mod_growth_policy<>(size)), std::length_error);
+    bucket_count = policy.max_bucket_count();
+    Policy policy2(bucket_count);
     
     
-    size = policy.max_bucket_count() + 1;
-    BOOST_CHECK_THROW((tsl::mod_growth_policy<>(size)), std::length_error);
+    bucket_count = std::numeric_limits<std::size_t>::max();
+    BOOST_CHECK_THROW((Policy(bucket_count)), std::length_error);
+    
+    
+    bucket_count = policy.max_bucket_count() + 1;
+    BOOST_CHECK_THROW((Policy(bucket_count)), std::length_error);
 }
